+++ src/main/java/nextstep/ladder/domain/user/LadderUsers.java
@@ -1,5 +1,6 @@
-package nextstep.ladder.domain.edge;
+package nextstep.ladder.domain.user;
 
+import nextstep.ladder.domain.prize.LadderPrizes;
 import nextstep.ladder.domain.line.Line;
 
 import java.util.List;
