+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -1,8 +1,12 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.edge.*;
 import nextstep.ladder.domain.line.Dimension;
-import nextstep.ladder.domain.line.LineFactory;
+import nextstep.ladder.domain.prize.LadderPrize;
+import nextstep.ladder.domain.prize.LadderPrizes;
+import nextstep.ladder.domain.user.LadderUser;
+import nextstep.ladder.domain.user.LadderUsers;
+import nextstep.ladder.domain.user.Position;
+import nextstep.ladder.factory.LineFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
