+++ src/test/java/nextstep/ladder/domain/edge/LadderPrizesTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.domain.edge;
 
+import nextstep.ladder.domain.prize.LadderPrize;
+import nextstep.ladder.domain.prize.LadderPrizes;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
