+++ src/test/java/nextstep/ladder/domain/edge/LadderResultTest.java
@@ -1,5 +1,10 @@
 package nextstep.ladder.domain.edge;
 
+import nextstep.ladder.domain.LadderResult;
+import nextstep.ladder.domain.prize.LadderPrize;
+import nextstep.ladder.domain.prize.LadderPrizes;
+import nextstep.ladder.domain.user.LadderUser;
+import nextstep.ladder.domain.user.LadderUsers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
