+++ src/test/java/nextstep/ladder/domain/edge/LadderUserTest.java
@@ -1,11 +1,11 @@
 package nextstep.ladder.domain.edge;
 
 import nextstep.ladder.domain.line.*;
+import nextstep.ladder.domain.user.LadderUser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
-import java.util.Random;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
