+++ src/test/java/nextstep/ladder/domain/line/PointFactoryTest.java
@@ -1,5 +1,6 @@
 package nextstep.ladder.domain.line;
 
+import nextstep.ladder.factory.PointFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
